+++ pkg/cmd/get/cmd.go
@@ -5,6 +5,7 @@ import (
 	"github.com/open-cluster-management/cm-cli/pkg/cmd/get/clusterclaim"
 	"github.com/open-cluster-management/cm-cli/pkg/cmd/get/clusterclaims"
 	"github.com/open-cluster-management/cm-cli/pkg/cmd/get/clusterpoolhosts"
+	"github.com/open-cluster-management/cm-cli/pkg/cmd/get/clusterpools"
 	"github.com/open-cluster-management/cm-cli/pkg/cmd/get/clusters"
 	"github.com/open-cluster-management/cm-cli/pkg/cmd/get/machinepools"
 	genericclioptionscm "github.com/open-cluster-management/cm-cli/pkg/genericclioptions"
@@ -28,6 +29,7 @@ func NewCmd(clusteradmFlags *genericclioptionsclusteradm.ClusteradmFlags, cmFlag
 	cmd.AddCommand(clusterpoolhosts.NewCmd(cmFlags, streams))
 	cmd.AddCommand(clusterclaims.NewCmd(cmFlags, streams))
 	cmd.AddCommand(clusterclaim.NewCmd(cmFlags, streams))
+	cmd.AddCommand(clusterpools.NewCmd(cmFlags, streams))
 
 	return cmd
 }