+++ pkg/cmd/scale/clusterpool/cmd.go
@@ -1,0 +1,51 @@
+// Copyright Contributors to the Open Cluster Management project
+package clusterpool
+
+import (
+	"fmt"
+
+	clusteradmhelpers "open-cluster-management.io/clusteradm/pkg/helpers"
+
+	"github.com/open-cluster-management/cm-cli/pkg/clusterpoolhost"
+	genericclioptionscm "github.com/open-cluster-management/cm-cli/pkg/genericclioptions"
+	"github.com/open-cluster-management/cm-cli/pkg/helpers"
+	cmdutil "k8s.io/kubectl/pkg/cmd/util"
+
+	"github.com/spf13/cobra"
+	"k8s.io/cli-runtime/pkg/genericclioptions"
+)
+
+var example = `
+# Scale clusterpool
+%[1]s scale clusterpool|cc <clusterpool_name> --size <size> <options>
+
+# Scale clusterpool on a given clusterpoolhost
+%[1]s scale clusterpool|cc <clusterpool_name> --size <size> <clusterpoolhost> <options>
+`
+
+// NewCmd ...
+func NewCmd(cmFlags *genericclioptionscm.CMFlags, streams genericclioptions.IOStreams) *cobra.Command {
+	o := newOptions(cmFlags, streams)
+	cmd := &cobra.Command{
+		Use:          "clusterpool",
+		Aliases:      []string{"cp"},
+		Short:        "Scale clusterpool",
+		Example:      fmt.Sprintf(example, helpers.GetExampleHeader()),
+		SilenceUsage: true,
+		PreRunE: func(cmd *cobra.Command, args []string) error {
+			clusteradmhelpers.DryRunMessage(cmFlags.DryRun)
+			return clusterpoolhost.BackupCurrentContexts()
+		},
+		Run: func(cmd *cobra.Command, args []string) {
+			cmdutil.CheckErr(o.complete(cmd, args))
+			cmdutil.CheckErr(o.validate())
+			cmdutil.CheckErr(o.run())
+		},
+		PostRunE: func(cmd *cobra.Command, args []string) error {
+			return clusterpoolhost.RestoreCurrentContexts()
+		},
+	}
+
+	cmd.Flags().Int32Var(&o.Size, "size", 1, "Set the size of a clusterpool")
+	return cmd
+}